requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.4.1
asyncpg==0.29.0
autopep8==1.5.7
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.2
click==8.1.7
config==0.5.1
cryptography==42.0.8
databases==0.9.0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jwt==1.3.1
libusb1==3.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ndeflib==0.3.3
nfcpy==1.0.4
npm==0.1.1
optional-django==0.1.0
orjson==3.10.3
psycopg2-binary==2.9.10
pycodestyle==2.7.0
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.18.2
pyDes==2.0.1
Pygments==2.18.0
PyMySQL==1.1.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-decouple==3.4
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2021.1
PyYAML==6.0.1
requests==2.32.4
rich==13.7.1
setuptools==69.5.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.4.1
starlette==0.37.2
toml==0.10.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
Unipath==1.1
urllib3==2.5.0
uuid==1.30
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wheel==0.43.0
whitenoise==5.3.0
Whoosh==2.7.4
